Add left/right keyboard movement
the movement doesn't feel great right now, but i'm going to wait
to tweak it until more gameplay is implemented.

Warning: friction implementation is mixed in with the movement keys.
and it works directly on velocity.
This works, but it does not use time at all, so changing framerate
will affect friction.
I'll circle back to that later.