Make ball bounce off player
The player's velocity is twiddled a little bit to point across the
court, and then added to the ball's velocity to get the final
velocity.
I like the way this feels a lot more than using vanilla elastic collision.
It's very forgiving, but it feels like you have a lot of control.