Show each player's controls in control selector
I think the controls section is finished for now.
The final version of the it looks like this:
     ________________
    /   Controls    /
   /---------------/
  / AI  /   W     /
 /     /  A   D  /
/-----/---------/

If the player is AI controlled, then the AI section will be highlighted
If the player is keyboard controlled, then the section with their movement
keys is highlighted.
Clicking on the bottom half of the widget switches between AI and keyboard
control.