Remove player air friction
Removing friction while in the air changes the feel of jumping a lot,
even though it doesn't change the player's ability to jump much.
Previously, friction was not applied when horizontal movement keys were
held down, so you could get the same effect by just holding the keys
in the air.
The main difference is that falling in a straight line is essentially
impossible now.