Re-use wall collision logic for ball/net collision
I just alter the ball's leftWallX and rightWallX properties depending
on which side of the screen it's on and whether it's above the net.

I was about to commit entirely new ball/net collision logic that
prevented the ball from punching through the net at high speeds
when i thought of this.
Not sure why I didn't think of this sooner, it's what I've been
doing for the players, and it makes the logic much simpler.