Handle non-alphanumeric keycodes slightly better
Instead of returning whatever random Unicode symbol corresponds to
the keycode, we now just return the string version of the key code.
Seeing raw key code numbers in the UI is not great UX, but it's
better than the previous random symbols, and handling it better than
this will require a fair amount of effort.

There are a few Elm packages that can map keycodes to a union type
with all the possible keys, which would make this slightly easier,
but none that, eg, map backspace, space, arrows, etc to strings.

If I improve on this in the future, it will likely be with either
SwiftsNamesake/proper-keyboard or ohanhi/keyboard-extra