Make arms move more independently
I decided to use the distance from the resting position to determine
if the ball is close enough to interact with instead of the shoulder
position. The shoulders are pretty close together, so you get both
arms essentially doing the same thing. With this change, it's more
common to get just one arm reaching for the ball.