Make ball bounce off player hands instead of body
This just makes the game a hell of a lot more ridiculous looking.
It's surprising how complex the arm movements get sometimes given
the extremely simple rules governing them.
The ball picks up velocity when it's in contact with the hands, and
the hands follow the ball, so it looks like the ball is being grabbed
and thrown.