Make AI chill out a bit
Instead of always relentlessly chasing the ball and jumping,
the new AI will only move toward the ball once its distance
reaches a threshold. Also, it will only jump if the ball is
close horizontally, and in the upper third of the screen.
I think this looks more natural, and works better with the new
arm mechanics. The old AI would try to center itself, which would
allow the ball to fall right between its hands.