Add instructions to instructions page
This page contains some basic text about gameplay and UI.
The lack of text wrapping in SVG made this fairly tedious.
The simplest way I've found to do this so far is to use
multiline strings for text, which you manually wrap (gq
works ok in Vim), then programmatically split into lines
and map to SVG tspan elements.
I'm using a variable width font, so a bit of trial and
error was required to get the wrapping right. If I have
to do much more of this, i'm just going to go monospace
and automate all of it.