Keep arm length constant
The arm would occasionally shrink/grow abruptly previously, and arms
being huge for a split second looks pretty bad.

In early iterations of the logic for this, I ran into some division
by zero/sqrt(-1) floating point errors.
The math for the final version should never produce conditions
like that, but I left some of the logic to check for and prevent
them because runtime errors are embarassing.