Add shockingly good explosion animation
I was going to replace the previous explosion animation with
a radial gradient because it was causing huge framerate drops
on other people's computers. Just for the hell of it, I applied
the old noise/displacement map filter to the new radial gradient,
and, WOW, it looks too good to not use. I'll stick a setting in
to disable it for slow computers.
I'll probably put a separate demo of just the explosion animation up
on GitHub, because it's wild how good it looks for 14 lines of SVG.