Stop key selection after player presses a key
Previously, when a player clicked on one of the keys in the
controls menu, that key would remain selected and we would keep
listening for new key presses until something else was clicked.
Now, it sets itself to the first key you press, and then is deselected.
I think that feels  more natural.