Use elm-lang/core/Color for colors in View
Just noticed that there's a color package. I converted all
the string colors in view code to Color.Color. This makes
the code a bit more verbose, but the type signatures are more
expressive and type checking will be more able to save me
from myself.
The Color module seems like it was designed for the old Collage
graphics package. There's no function for converting colors
to strings, so I grabbed another package for that.
There's no package that converts the gradient types in Color
to anything that would fit in HTML or SVG, so I'm not using
them.
Both of those conversions would be really straightforward to
write myself, but the cost:benefit is just not there for this
project.