Move controls to its own page
The new page has a lot more room than the old popout thing on
the main menu, so I can add more options to it.
This allows me to get rid of all of the "sub menu" stuff.
Note: I just slapped the old controls menu on a new page without
changing it at all so it looks REALLY bad right now. It works though.