Redo controls menu with new Grid module
The controls menu (and most of the rest of my UI) is built from
parallelograms. I was really not happy with the code that was
reponsible for this. It was not very flexible and relied on a ton of
little bits of math that were spread out all over the place. I also
had the key realization that I could just treat everything like it was
rectangular while putting the UI together, and then at the very I can
shift all the x values at once and get the same result.

I created the Grid module to pull most of that logic out of the View
and handle it all in a cleaner way. It should work for the main menu
and the HUD as well.