Fix drawRegion text size calculation
It was using Grid.rowHeight, which is only helpful for getting the size
of a single row without constructing an entire Region.
Grid.regionHeight works on the whole region and accounts for padding
correctly. I also changed the height multiplier from 0.7 to 5/6
because I think it looks a little better.