Add name field to Player record
I haven't used this new field yet, but having the name available
in the player record should make some view logic simpler.
The main motivation is for the HUD refactoring i'm doing. All
the other hud information comes straight out of the players and
i prefer this to needing to hardcode it again for that.