Handle side in drawRegion
This is necessary to use it for the HUD. The plan is to split
drawRegion into 2 functions later today: one for text, and one for colored
background polygons, so this is a pretty quick and dirty
change so I could focus on the HUD for a bit.

One key thing that is embarassing and MUST BE changed is that I
hardcoded the screen width in drawRegion. Passing that down would
require me to modify the signatures of functions that call drawRegion
and, since I'm about to drastically rework that anyway, I didn't want
to go to the effort.

In retrospect, those changes would have been fewer keystrokes than
this explanation though, so I probably should have just done it