Use Grid for control toggle
This was the last place drawUiBlock was used, so I got to delete that
and all of its dependencies, which was incredibly satisfying.

One behavioral change: the control toggle example on the help page
is active and will cause model.player1.ai to toggle if you click it
I thought about preventing this with Html.map, but I kind of like it
the way it is.

Now that everything is gridded, I can move on to the next major
refactor: separating text and shape rendering.

This is roughly the pipeline I'm going to shoot for:

           grid
       ______|____
      /           \
     /             \
transform         transform
to polygon         to text
  paths           positions
    |                 |
account for       account for
  side              side
    |                 |
transform to      transform to
List (Svg Msg)    List (Svg Msg)
     \               /
      \_____________/
             |
          concat
             |
        add to view

(why yes I *am* on an ASCII art kick right now, how could you tell?)

The inner 3 steps are pretty different for text and polygons, but
they're all lumped together right now. This causes functions to have
A LOT more branching than they need to. Splitting it up should
take care of that.