Unify UI grid handling
This is a pretty big set of UI changes and deviates a little bit from
my original plan.
Here's how things work:
- there is a union type for every kind of UI element
- everything builds grids of these
there's a UiPrimitive type that describes polygons and text
the uiElementToPrimitives function translates all of these different
UI elements to primitives, and then the primitives are drawn.

Overall this is a huge simplification of the pipeline, but there are
some rough edges:
- uiElementToPrimitives is HUGE. It is very straightforward though, it
  just handles all the possibilities for a large union type. I will
  split this up into some helper functions.
- There are a few places with nested list transformations that aren't
  super readable.
- There is no semblance of logic behind the ordering of the view
  functions.
- The hud control toggle does not show the correct controls.
- the pause menu duplicates some of the grid logic