Use same message for toggling either player ai
Replaced:
TogglePlayer1Ai, TogglePlayerAi
with
ToggleAi Left, ToggleAi Right

I added a mapPlayer function that makes it easy to modify
a player in the model based on the side.

Combined, these allowed me to get rid of most of the
player1 <=> left, player2 <=> right case statements

I feel just a little weird about using side to refer to players,
but it is unambiguous and things that are graphically on the left
are typically relevant to the player that's on the left too, so it's
a useful equivalence to have.

I added the alias PlayerSide to make it a little clearer what
is being referred to.