Handle ball in middle of net
I've had this case set up to crash whenever it happened for a while
now and I've never observed it except when I intentionally modified
the code to trigger it.

This tells me that it really doesn't matter much how it's handled, so
I just defaulted it to acting like it's on the right side of the net.